Ensure that the archive doesn't list pages the user can't access
This makes sure that the cache for archive pages is never used when a
page is listed that the current user can't access. For simplicity and
speed reasons the opposite check, i.e. if pages are missing that the
current user could access, is missing. It is assumed that this isn't the
typical use case for this plugin, a simple workaround is using the
~~NOCACHE~~ macro in the page.